Decimal double fuckery

+++ ReykjanesRidge/ReykjanesRidge.Server/Shared/FomRangeSliderEventArgs.cs
@@ -2,8 +2,8 @@
 {
     public class FomRangeSliderEventArgs : EventArgs
     {
-        public double Range { get; set; }
-        public double FirstValue { get; set; }
-        public double SecondValue { get; set; }
+        public decimal Range { get; set; }
+        public decimal FirstValue { get; set; }
+        public decimal SecondValue { get; set; }
     }
 }
